Use `SetCurrentDirectoryToEmptyTemporaryDirectory` in first test classes
