+++ Tests/Testably.Abstractions.Tests/FileSystemDirectoryInfoFactoryTests.cs
@@ -7,19 +7,17 @@ public abstract class FileSystemDirectoryInfoFactoryTests<TFileSystem>
 {
     #region Test Setup
 
-    public string BasePath { get; }
+    public abstract string BasePath { get; }
 
     public TFileSystem FileSystem { get; }
     public ITimeSystem TimeSystem { get; }
 
     protected FileSystemDirectoryInfoFactoryTests(
         TFileSystem fileSystem,
-        ITimeSystem timeSystem,
-        string basePath)
+        ITimeSystem timeSystem)
     {
         FileSystem = fileSystem;
         TimeSystem = timeSystem;
-        BasePath = basePath;
     }
 
     #endregion
