+++ Tests/Testably.Abstractions.Tests/Mock/MockFileSystem.DirectoryInfoFactoryTests.cs
@@ -9,19 +9,21 @@ public static partial class MockFileSystem
     public sealed class
         DirectoryInfoFactoryTests : FileSystemDirectoryInfoFactoryTests<FileSystemMock>
     {
+        /// <inheritdoc cref="FileSystemDirectoryInfoFactoryTests{TFileSystem}.BasePath" />
+        public override string BasePath => _directoryCleaner.BasePath;
+
+        private readonly FileSystemInitializer.IDirectoryCleaner _directoryCleaner;
+
         public DirectoryInfoFactoryTests() : this(new FileSystemMock())
         {
         }
 
         private DirectoryInfoFactoryTests(FileSystemMock fileSystemMock) : base(
             fileSystemMock,
-            fileSystemMock.TimeSystem,
-            fileSystemMock.Path.Combine(
-                fileSystemMock.Path.GetTempPath(),
-                fileSystemMock.Path.GetRandomFileName()))
+            fileSystemMock.TimeSystem)
         {
-            FileSystem.Directory.CreateDirectory(BasePath);
-            FileSystem.Directory.SetCurrentDirectory(BasePath);
+            _directoryCleaner = FileSystem
+               .SetCurrentDirectoryToEmptyTemporaryDirectory();
         }
     }
 }
